more path cleanup
