+++ tests/benchmark_models/evaluate_benchmark.py
@@ -9,8 +9,11 @@ import os
 import matplotlib.pyplot as plt
 import pandas as pd
 import seaborn as sns
+from pathlib import Path
 
 # read benchmark results for different models
+
+script_dir = Path(__file__).parent.resolve()
 
 outfile = "computation_times.csv"
 df = pd.concat(
@@ -18,7 +21,7 @@ df = pd.concat(
         pd.read_csv(f, header=[0], index_col=[0])
         .rename(columns={"0": "_".join(f.split("_")[:2])})
         .T
-        for f in os.listdir()
+        for f in os.listdir(script_dir)
         if f.endswith(".csv")
         if f != outfile
     ]
